pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agente-negotiator"
version = "0.1.0"
description = "MVP – Agente negociador de influenciadores"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.21.0",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "sqlalchemy>=2.0",
    "typer[all]>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
negotiator = "app.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
